requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
pydantic>=2.7.4
pydantic-settings>=2.3.4
pandas>=2.2.2
supabase>=2.5.1
pyjwt>=2.8.0
requests>=2.32.3
pdfplumber>=0.11.0
google-generativeai>=0.7.1
google-auth>=2.29.0
Pillow>=10.3.0
pypdfium2>=4.28.0
rich>=13.7.1
python-dotenv>=1.0.1
